backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
uuid==1.30
